Fix READMEs and providers to allow passing `peer` to examples as a URL param
In many of the examples published, the README contains a reference to passing a custom sync server. These references refer variously to a `peer` or a `sync` URL parameter. This commit unifies all these, consolidating on a single `peer` param, which applies across all examples where this is referenced in the README.