Make the whole surface area of the button clickable
Currently, the Button component wraps a Next Link. That means the
Button's padding is applied around the link, so only the text in the
button is clickable. This changes the order of the components so that
the Button is fully wrapped by the Link, meaning the whole of the button
is now clickable.